fixed some bugs in eval and postprocess sweep. Improved to an F1 of 0.8! With postprocess sweep, I think we can improve that number.

+++ torch_mesmer/eval.py
@@ -138,7 +138,7 @@ def main(device: str,
     }
 
     X_test = z_test['X'][:]
-    y_test = z_test['y'][:]
+    y_test = z_test['y'][:].astype(int)
     y_test = np.flip(y_test, axis=1)
 
     mpps = z_test['mpp'][:]
